requirements.txt
discord.py==2.3.2
python-dotenv==1.0.1
audioop-lts==0.2.0
PyNaCl==1.6.1
pymongo==4.6.0